As expected, the second day fuckin'-
For real this time, it's the second day of something I thought was impossible to solve but was met with a surprising solution that took me just–aughh god damn that was terrible. Well at least I'm working on Jelly Jamm Venice Season 3 Episode 8 still to distract myself from it.

Btw the last thing I have to do now is to convert a `String` to a `const char*`. That's for later.